docs(davinci): record p2-11 publish graph firewall
