+++ davinci-road/plan/phase-2-records/p2-11.md
@@ -7,15 +7,17 @@ in `phase-2.md` stays **unticked** until the series completes the
 acceptance gate. Do not invent installments that have not merged to
 `origin/main`.
 
-The unpublished home is `vize_ricalco::emit`. `vize_atelier_dom` is
-published and cannot name Davinci crates in its release graph
-(phase-2.md: the program decision is still open). Dual-run lives in
-atelier_dom **test space** (dev-dep carve-out). The old DOM lane is
-still the shipped compile path. The exit gate is unchanged.
+The S1-to-S2 emit test-consumer import path is `vize_s1_to_s2::emit`;
+the physical package still uses the codename `vize_ricalco` until a
+dedicated naming slice. `vize_atelier_dom` is published and cannot name
+unpublished Davinci crates in its release graph. The current decision is a
+firewall: S2 DOM witnesses live in atelier_dom **test space** through
+dev-dependencies that are stripped on publish, and the tooling gate rejects any
+unstripped edge from a publishable crate into an unpublished stage crate. The
+old DOM lane is still the shipped compile path. The exit gate is unchanged.
 
 The `emit.rs` module docs keep a running capability list. **This
-record's landed numbers are the thirty-nine `origin/main` PRs**, not that
-comment.
+record's landed numbers are the forty `origin/main` PRs**, not that comment.
 
 ## Installment table
 
@@ -60,6 +62,7 @@ comment.
 | 37  | 2026-08-29 | [#5208](https://github.com/ubugeeei-prod/vize/pull/5208) | `4e577b62a` | emit object `v-bind` modifiers               |
 | 38  | 2026-08-29 | [#5210](https://github.com/ubugeeei-prod/vize/pull/5210) | `f3959e7e3` | emit object `v-on` modifiers                 |
 | 39  | 2026-08-29 | [#5212](https://github.com/ubugeeei-prod/vize/pull/5212) | `22674520f` | cover recent patch-flag witnesses            |
+| 40  | 2026-08-29 | [#5214](https://github.com/ubugeeei-prod/vize/pull/5214) | `be344e787` | pin publish graph firewall                   |
 
 `landed` is the UTC committer date on `origin/main` (the merge-to-main
 clock). Per-installment records live in [p2-11/](./p2-11/):
@@ -103,8 +106,9 @@ clock). Per-installment records live in [p2-11/](./p2-11/):
 - [Installment 37 — object `v-bind` modifiers](./p2-11/installment-37.md)
 - [Installment 38 — object `v-on` modifiers](./p2-11/installment-38.md)
 - [Installment 39 — recent patch-flag witness](./p2-11/installment-39.md)
+- [Installment 40 — publish graph firewall](./p2-11/installment-40.md)
 
-## Current named remainder (after #5212)
+## Current named remainder (after #5214)
 
 Installment 20 closes the dynamic-argument `v-bind` key/modifier item that the
 installment-19 snapshot correctly left open. The current `emit.rs` capability
@@ -190,12 +194,17 @@ and object-spread surfaces: `v-show`, `v-html`, `v-text`, `v-cloak`, object
 `v-bind` modifiers and object `v-on` modifiers now share a dedicated
 S2-vs-shipped per-node patch-flag test. This is a witness expansion, not a
 production-lane switch.
+Installment 40 pins the current publish graph decision. The stage crates remain
+unpublished in this phase, but publishable crates may exercise them only through
+dev-dependencies that Cargo strips on publish. A normal or build dependency from
+a publishable crate into `vize_davinci`, `vize_s1`, `vize_s2` or
+`vize_ricalco` is now a tooling failure, so the old DOM lane cannot be replaced
+by an accidental unresolvable release graph edge.
 
 The series is not done. Waiver budget remains zero. The old DOM lane remains
-the shipped path; the publish-graph decision, hydrated full-corpus comparison
-count, patch-flag equivalence and DOM allocation budget are still task-level
-blockers. The flag's deletion is still an exit-gate line, not this record's to
-tick.
+the shipped path; the hydrated full-corpus comparison count, patch-flag
+equivalence and DOM allocation budget are still task-level blockers. The flag's
+deletion is still an exit-gate line, not this record's to tick.
 
 ## Not series installments
