+++ tests/tooling/davinci-stage-dependencies.test.ts
@@ -1,126 +1,19 @@
 import assert from "node:assert/strict";
-import { spawnSync } from "node:child_process";
 import fs from "node:fs";
 import path from "node:path";
 import { test } from "node:test";
-import { fileURLToPath } from "node:url";
 
-const repoRoot = path.resolve(path.dirname(fileURLToPath(import.meta.url)), "../..");
-
-type Dependency = {
-  name: string;
-  rename: string | null;
-  kind: "dev" | "build" | null;
-  req: string;
-};
-
-type Package = {
-  name: string;
-  dependencies: Dependency[];
-  manifest_path: string;
-  publish: string[] | null;
-};
-
-type Metadata = { packages: Package[] };
-
-function readMetadata(): Metadata {
-  const result = spawnSync(
-    "cargo",
-    ["metadata", "--no-deps", "--format-version", "1", "--locked"],
-    { cwd: repoRoot, encoding: "utf8", maxBuffer: 64 * 1024 * 1024 },
-  );
-  assert.equal(result.status, 0, result.stderr);
-  return JSON.parse(result.stdout) as Metadata;
-}
-
-function workspacePackage(metadata: Metadata, name: string): Package {
-  const found = metadata.packages.find((pkg) => pkg.name === name);
-  assert.ok(found, `workspace package ${name} must exist`);
-  return found;
-}
-
-function dependency(
-  metadata: Metadata,
-  packageName: string,
-  dependencyName: string,
-  kind: Dependency["kind"],
-): Dependency {
-  const found = workspacePackage(metadata, packageName).dependencies.find(
-    (dependency) => dependency.kind === kind && dependency.name === dependencyName,
-  );
-  assert.ok(found, `${packageName} must declare ${kind ?? "normal"} dependency ${dependencyName}`);
-  return found;
-}
-
-function readRepoFile(...segments: string[]): string {
-  return fs.readFileSync(path.join(repoRoot, ...segments), "utf8");
-}
-
-function* walkRustFiles(directory: string): Generator<string> {
-  for (const entry of fs.readdirSync(directory, { withFileTypes: true })) {
-    const fullPath = path.join(directory, entry.name);
-    if (entry.isDirectory()) {
-      yield* walkRustFiles(fullPath);
-    } else if (entry.isFile() && entry.name.endsWith(".rs")) {
-      yield fullPath;
-    }
-  }
-}
-
-function s2DomWitnessFiles(): string[] {
-  const testDir = path.join(repoRoot, "crates", "vize_atelier_dom", "tests");
-  const witnesses = fs
-    .readdirSync(testDir, { withFileTypes: true })
-    .filter((entry) => entry.isFile() && /^davinci_s2_.*\.rs$/u.test(entry.name))
-    .map((entry) => entry.name)
-    .sort();
-  assert.ok(
-    witnesses.length >= 20,
-    `expected broad S2 DOM witness coverage, found only ${witnesses.length} files`,
-  );
-  return [...witnesses, path.join("support", "mod.rs")];
-}
-
-function assertS0AliasConsumer(options: {
-  packageName: string;
-  label: string;
-  directory: string;
-  filter?: (fullPath: string) => boolean;
-}) {
-  const { packageName, label, directory, filter = () => true } = options;
-  const dependencies = workspacePackage(metadata, packageName).dependencies;
-  assert.ok(
-    dependencies.some(
-      (dependency) =>
-        dependency.kind === null &&
-        dependency.name === "vize_carton" &&
-        dependency.rename === "vize_s0",
-    ),
-    `${packageName} must import vize_carton as vize_s0 for S0 storage`,
-  );
-  assert.ok(
-    dependencies.every(
-      (dependency) => dependency.name !== "vize_carton" || dependency.rename === "vize_s0",
-    ),
-    `${packageName} must not depend on vize_carton through its physical name`,
-  );
-
-  const offenders = [];
-  let aliasImports = 0;
-  for (const fullPath of walkRustFiles(directory)) {
-    if (!filter(fullPath)) continue;
-    const source = fs.readFileSync(fullPath, "utf8");
-    if (/\bvize_carton::|use vize_carton\b/u.test(source)) {
-      offenders.push(path.relative(repoRoot, fullPath));
-    }
-    if (/\bvize_s0::|use vize_s0\b/u.test(source)) {
-      aliasImports += 1;
-    }
-  }
-
-  assert.ok(aliasImports > 0, `${label} should use the vize_s0 alias`);
-  assert.deepEqual(offenders, []);
-}
+import {
+  assertS0AliasConsumer,
+  dependency,
+  metadata,
+  readRepoFile,
+  repoRoot,
+  s2DomWitnessFiles,
+  walkRustFiles,
+  workspacePackage,
+  type Package,
+} from "./support/davinci-stage-dependencies.ts";
 
 const aliases = new Map<string, ReadonlyArray<readonly [string, string | null]>>([
   ["vize_davinci", [["vize_carton", "vize_s0"]]],
@@ -135,8 +28,6 @@ const aliases = new Map<string, ReadonlyArray<readonly [string, string | null]>>
     ],
   ],
 ]);
-
-const metadata = readMetadata();
 
 const unpublishedDavinciStages = new Set(["vize_davinci", "vize_s1", "vize_s2", "vize_ricalco"]);
